requirements.txt
atlassian-python-api<4
ipdb 
python-dotenv<0.20